Update files to v1.15.2

+++ Systems/WorldGen/Standard/ChunkGen/1.GenTerra/GenTerra.cs
@@ -328,12 +328,12 @@ namespace Vintagestory.ServerMods
                 }
             }
 
-            int ymax = 0;
+            ushort ymax = 0;
             for (int i = 0; i < rainheightmap.Length; i++)
             {
                 ymax = Math.Max(ymax, rainheightmap[i]);
             }
-            chunks[0].MapChunk.YMax = (ushort)ymax;
+            chunks[0].MapChunk.YMax = ymax;
         }
